Fix unsigned overflow in SPath::get negative-index loop
i += vPlots.size() with i=-1 (int) and vPlots.size()=9 (size_t):
  implicit conversion: (unsigned int)(-1) = 4294967295
  4294967295 + 9 = 4294967304 > UINT_MAX  -> UBSan fires

Result after wrap is 8 (correct: last of 9 plots), but the
arithmetic is unintentionally unsigned.  The intent is signed:
  -1 + (int)9 = 8  no conversion, no overflow.

Cast vPlots.size() to int so the while-loop body stays in
signed arithmetic.  All five get(-1) callers in CvMilitaryAI.cpp
(UpdateAttackTargets land/water paths, SetWaypoints target index)
route through this single function.

Triggered: Germany (player 1) DoTurn -> UpdateAttackTargets,
land path of 9 plots, get(-1) fetching enemy city owner.

+++ CvGameCoreDLL_Expansion2/CvAStar.cpp
@@ -4185,7 +4185,7 @@ CvPlot * SPath::get(int i) const
 
 	//allow negative indices
 	while (i < 0)
-		i += vPlots.size();
+		i += (int)vPlots.size();
 
 	return GC.getMap().plotUnchecked(vPlots[i].x,vPlots[i].y);
 }
